Add more encapsulation mechanism
When a unit test does not have prologue code and an assertion
fails, an internal assertion fails in the call to
unitTestCounter.current.revertPassing() caused by the
UnitTestCountertruct invariant.

This is solved using a flag to determine whether the current
execution occurred inside a unit test block containing a prologue
or not.