Pass unit test configuration file as command-line argument
Acquiring the contents of the unit test configuration file
using import statement could not check whether the file exists
or not. The file is therefore required to exist even though it
does not need to be.

This implementation of using unit test configuration file to get
the list of unit test blocks and modules to execute is better.
The file could be checked for existence and multiple files could be
passed via the command-line.